match: three referent/signedness bugs in Io + Utils (SetCurLevel -> 100 EXACT)
CSaveGame::SetCurLevel 0xe5660 94.00 -> 100.00 EXACT. The guard was `v <=
CurrentLevel()`, whose accessor casts the u32 storage back to the signed
QuestLevel enum, so cl emitted `jle` where retail has `jbe`. Comparing `v` against
the raw `m_curLevel` member restores the unsigned compare by the usual arithmetic
conversions - no cast needed. The asymmetry in retail is the tell: the FIRST guard
compares against an enum CONSTANT and is signed (`jge`, matched all along), the
second compares against the u32 MEMBER and is unsigned.

operator<<(ostream&, const zBitVec&) 0x193080: retail calls
`??6ostream@@QAEAAV0@I@Z` (unsigned int) where we called the `H` (int) overload -
`static_cast<i32>(i)` on an already-unsigned loop index. Dropping the cast picks
the right overload; a 9-cell loop-type x streamed-type Cartesian confirms the
unsigned index is also what retail's loop uses.

GetGruntzDriveLetter 0x1ffe0: the cached drive letter was a function-local
`static char s_cdDriveLetter` AND a placeholder global `g_val_22b25c[1]` at
0x22b25c sitting one line above it. Retail references 0x22b25c at exactly the two
sites we referenced the static, so they are the same datum - homed as
`char g_cdDriveLetter`, which also retires a `g_<hex>` placeholder name.

+++ src/Utils/BitArrayStream.cpp
@@ -16,7 +16,7 @@ ostream& operator<<(ostream& accum, const zBitVec& bits) {
                 if (!first) {
                     accum << ' ';
                 }
-                accum << static_cast<i32>(i);
+                accum << i;
                 first = 0;
             }
         }
